requirements.txt
# Core Django Framework
django>=5.0,<6.0

# Django REST Framework for API development
djangorestframework>=3.15,<4.0

# JWT Authentication for APIs
djangorestframework-simplejwt>=5.3,<6.0

# Allauth for registration, login, and social authentication
django-allauth>=0.63,<1.0

# Rate limiting to prevent abuse
django-ratelimit>=4.1,<5.0

# Handle Cross-Origin requests from frontend
django-cors-headers>=4.3,<5.0

# Optional: environment variable management
python-dotenv>=1.0,<2.0

requests

cryptography

twilio

dj_rest_auth